+++ docs/guide/reset.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './reset.css';
+
+const Demo = () => 'haha';
+
+export default Demo;