+++ test/index-spec.js
@@ -1,8 +1,0 @@
-import expect from 'expect.js';
-import Component from '../index';
-
-describe('index.js', () => {
-  it('should export.', () => {
-    expect(Component).to.not.eql(undefined);
-  });
-});